% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cjniv3}
\alias{cjniv3}
\title{Codes et libelle des catégories juridiques de niveau 3}
\format{
A data frame with x rows and y variables:
\describe{
\item{Code}{les codes des CJ niv3}
\item{Libelle}{les libelles des CJ niv3}
}
}
\usage{
cjniv3
}
\description{
A dataset contenant les codes et libelle des CJ niveau 3
Les variables sont les suivantes
}
\keyword{datasets}
