% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneutils.R
\name{findmycj}
\alias{findmycj}
\title{findmycj}
\usage{
findmycj(cj)
}
\arguments{
\item{cj}{une chaine de caracteres}
}
\value{
Une chaine de caracteres.
}
\description{
Retourne le libelle d'un code CJ de niveau 3.
}
\examples{
\dontrun{
findmycj("5599")
}
}
