% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneutils.R
\name{findmynaflib}
\alias{findmynaflib}
\title{findmynaflib}
\usage{
findmynaflib(nafc)
}
\arguments{
\item{nafc}{une chaine de caracteres}
}
\value{
Une chaine de caracteres.
}
\description{
Retourne le libelle d'un code naf 5 caracteres.
}
\examples{
\dontrun{
findmynaflib("01.21Z")
}
}
