% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneutils.R
\name{findmytreff}
\alias{findmytreff}
\title{findmytreff}
\usage{
findmytreff(tr)
}
\arguments{
\item{tr}{une chaine de caracteres}
}
\value{
Un tibble ou rien si pas d'echo
}
\description{
Retourne le libelle d'un code tranche d'effectif
}
\examples{
findmytreff("42")
}
