% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneutils.R
\name{get_api_key}
\alias{get_api_key}
\title{get_api_key}
\usage{
get_api_key(keyname = "")
}
\arguments{
\item{keyname}{le nom de la variable d'environnement contenant votre clef de l'API Sirene.}
}
\value{
Une chaine de caracteres.
}
\description{
Retourne la valeur de la variable d'environnement INSEESIRENE_APIKEY
}
\details{
Pour l'instant, la valeur de la variable d'environnement est fixee a INSEESIRENE_APIKEY.
}
\examples{
\dontrun{
get_api_key("INSEESIRENE_APIKEY")
}
}
