% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneget.R
\name{insee_get_liens_succession}
\alias{insee_get_liens_succession}
\title{insee_get_liens_succession}
\usage{
insee_get_liens_succession(type_lien = "S", etab = NULL)
}
\arguments{
\item{type_lien}{le type de lien recherche : predecesseur (P) ou successeur (S)}

\item{etab}{le numero siret de l'etablissement dont on cherche le lien de succession}
}
\value{
un data.frame des donnees renvoyees par l'API Sirene
}
\description{
Retourne le resultat de la requete liens de succession
}
\examples{
\dontrun{
insee_get_liens_succession(type_lien = "S", etab = "123 456 789 00012")
}
}
