% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneget.R
\name{insee_get_query}
\alias{insee_get_query}
\title{insee_get_query}
\usage{
insee_get_query(url, simp = TRUE)
}
\arguments{
\item{url}{l'url a interroger}

\item{simp}{if TRUE (default), simplifyVector}
}
\value{
un data.frame des donnees renvoyees par l'API Sirene
}
\description{
Retourne le resultat de la requete
}
\examples{
\dontrun{
insee_get_query("https://api.insee.fr/api-sirene/3.11/siret/40188740100040")
}
}
