% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneget.R
\name{insee_get_siren}
\alias{insee_get_siren}
\title{insee_get_siren}
\usage{
insee_get_siren(siren = NULL, champs = NULL)
}
\arguments{
\item{siren}{le numero siren a requeter}

\item{champs}{limiter le champ de la reponse siren. Pas encore actif TODO}
}
\value{
un data.frame parse du resultat de la requete sur numero siren
}
\description{
Retourne le resultat de la requete sur un numero siren
}
\examples{
\dontrun{
insee_get_siren("monnumerosiren")
}
}
