% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneget.R
\name{insee_get_siret}
\alias{insee_get_siret}
\title{insee_get_siret}
\usage{
insee_get_siret(siret = NULL, simp = TRUE)
}
\arguments{
\item{siret}{le code siret a interroger}
}
\value{
un data.frame des donnees renvoyees par l'API Sirene
}
\description{
Retourne le resultat parse de la requete sur numero siret
}
\examples{
insee_get_siret(siret = "40188740100040")
}
