% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneget.R
\name{insee_get_siret_all}
\alias{insee_get_siret_all}
\title{insee_get_siret_all}
\usage{
insee_get_siret_all(siret = NULL)
}
\arguments{
\item{siret}{le code siret a interroger}
}
\value{
une liste non parsee retournee par l'API
}
\description{
Retourne le resultat (non parse) de la requete sur numero siret
}
\examples{
\dontrun{
insee_get_siret_all("40188740100040")
}
}
