% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneget.R
\name{insee_get_siret_multi_simp}
\alias{insee_get_siret_multi_simp}
\title{insee_get_siret_multi_simp}
\usage{
insee_get_siret_multi_simp(url)
}
\arguments{
\item{url}{l'url multicritères a adresser a l'API}
}
\value{
un data.frame parse du resultat de la requete multicritères
}
\description{
Retourne le resultat de la requete multicriètre sur numeros siret dans un autre format
}
\examples{
insee_get_siret_multi_simp("https://api.insee.fr/api-sirene/3.11/siret?q=codeCommuneEtablissement\%3A13*\%20AND\%20activitePrincipaleUniteLegale\%3A10.71*&nombre=1000&curseur=*")
}
