% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneparse.R
\name{insee_parse_siren_resp}
\alias{insee_parse_siren_resp}
\title{insee_parse_siren_resp}
\usage{
insee_parse_siren_resp(x)
}
\arguments{
\item{x}{une réponse siren mono unite legale}
}
\value{
un data.frame parse d'une requête siren mono unite legale
}
\description{
Parse le résultat d'une requête siren mono UL
}
\examples{
insee_parse_siren_resp(objetnonparsedunerequetesiren)
}
