% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneparse.R
\name{insee_parse_siret_multi}
\alias{insee_parse_siret_multi}
\title{insee_parse_siret_multi}
\usage{
insee_parse_siret_multi(x)
}
\arguments{
\item{x}{une reponse siret multicriteres}
}
\value{
un data.frame parse d'une requete multicriteres sur des etab siret
}
\description{
Parse le resultat d'une requete multicriteres de siret
}
\examples{
\dontrun{
insee_parse_siret_multi(objetnonparsedunerequetesiretmulticriteres)
}
}
