% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneparse.R
\name{insee_parse_siret_resp}
\alias{insee_parse_siret_resp}
\title{insee_parse_siret_resp}
\usage{
insee_parse_siret_resp(x)
}
\arguments{
\item{x}{une reponse siret sur un etablissement}
}
\value{
un data.frame parse d'une requete mono etab siret
}
\description{
Parse le resultat d'une requete sur UN numero siret
}
\examples{
\dontrun{
insee_parse_siret_resp(objetnonparsedunerequetesiretmonoetab)
}
}
