% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneparse.R
\name{insee_parse_siret_resps}
\alias{insee_parse_siret_resps}
\title{insee_parse_siret_resps}
\usage{
insee_parse_siret_resps(x)
}
\arguments{
\item{x}{une réponse siret sur plusieurs etablissements}
}
\value{
un data.frame parse d'une requête multi etab siret
}
\description{
Parse le résultat d'une requête sur un ensemble de siret
}
\examples{
insee_parse_siret_resps(objetnonparsedunerequetesiretmultietab)
}
