% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneutils.R
\name{insee_utils_avis_sit}
\alias{insee_utils_avis_sit}
\title{insee_utils_avis_sit}
\usage{
insee_utils_avis_sit(siret, url = FALSE)
}
\arguments{
\item{siret}{le siret a interroger}
}
\value{
un avis de situation sirene
}
\description{
Renvoie un avis de situation a partir d'un siret
}
\examples{
insee_utils_avis_sit(siret = "12345678900001")
}
