% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneutils.R
\name{insee_utils_concat_multi}
\alias{insee_utils_concat_multi}
\title{insee_utils_concat_multi}
\usage{
insee_utils_concat_multi(url = NULL, multi = "", nb = TRUE)
}
\arguments{
\item{url}{l'url a interroger}

\item{multi}{la chaine de caracteres des parametres de la requete multicriteres}

\item{nb}{booleen si vrai ajoute a l'url le nb max d'echos autorises par l'api (1000)}
}
\value{
une chaine de char
}
\description{
Retourne la chaine concatenee d'une requete multicriteres
}
\examples{
\dontrun{
insee_utils_concat_multi("https://api.insee.fr/api-sirene/3.11/siret?q=", multi = "")
}
}
