% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneutils.R
\name{insee_utils_find_nb_periodes}
\alias{insee_utils_find_nb_periodes}
\title{insee_utils_find_nb_periodes}
\usage{
insee_utils_find_nb_periodes(df)
}
\arguments{
\item{df}{le data.frame utilise}
}
\value{
une chaine de caractere : le nombre de periodes d'un etablissement
}
\description{
Retourne le nombre de periodes disponibles pour un etablissement
}
\examples{
\dontrun{
insee_utils_find_nb_periodes(mydf)
}
}
