% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneutils.R
\name{insee_utils_get_nic}
\alias{insee_utils_get_nic}
\title{insee_utils_get_nic}
\usage{
insee_utils_get_nic(x)
}
\arguments{
\item{x}{la requete sur siren a interroger}
}
\value{
une chaine de caractere : le code nic du siege de l'unite legale
}
\description{
Renvoie le nic du siege a partir d'une requete sur siren
}
\examples{
insee_utils_get_nic(x = marequetesiren)
}
