% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneutils.R
\name{insee_utils_howmany_echos}
\alias{insee_utils_howmany_echos}
\title{insee_utils_howmany_echos}
\usage{
insee_utils_howmany_echos(url)
}
\arguments{
\item{url}{l'url a interroger}
}
\value{
une chaine de caractere : le nombre d'echos
}
\description{
Retourne le nombre d'echos de la requete
}
\examples{
\dontrun{
insee_utils_howmany_echos("https://api.insee.fr/api-sirene/3.11/siret?q=codeCommuneEtablissement\%3A13*\%20AND\%20activitePrincipaleUniteLegale\%3A10.71*&nombre=1000&curseur=*")
}
}
