% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneutils.R
\name{insee_utils_make_report}
\alias{insee_utils_make_report}
\title{generer un rapport html d'un etablissement}
\usage{
insee_utils_make_report(siren)
}
\arguments{
\item{siret}{une chaine de caractere pour l'etablissement a consulter}
}
\value{
un fichier html, affiche dans le browser
}
\description{
Affiche une page html synthétisant de l'information de base sur un etablissement
}
\examples{
insee_utils_make_report(siret = "40188740100040")
}
