% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneutils.R
\name{insee_utils_remove_url_curseurx}
\alias{insee_utils_remove_url_curseurx}
\title{insee_utils_remove_url_curseurx}
\usage{
insee_utils_remove_url_curseurx(url)
}
\arguments{
\item{url}{l'url a modifier}
}
\value{
une chaine de caractere (l'url) sans le curseur
}
\description{
Retourne une url sans le fragment curseur=*
}
\examples{
insee_utils_remove_url_cruseurx("https://mapetiteurl?curseur=*")
}
