% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneutils.R
\name{insee_utils_resp_code}
\alias{insee_utils_resp_code}
\title{insee_utils_resp_code}
\usage{
insee_utils_resp_code(url, mode = "S")
}
\arguments{
\item{url}{l'url a interroger}

\item{mode}{le mode d'interrogation : simple (S) ou etendu (E)}
}
\value{
une chaine de caractere : le code retour de la requete
}
\description{
Renvoie le code retour de la requete
}
\examples{
\dontrun{
insee_utils_resp_code("https://api.insee.fr/api-sirene/3.11/siret?q=codeCommuneEtablissement\%3A13*\%20AND\%20activitePrincipaleUniteLegale\%3A10.71*&nombre=1000&curseur=*")
}
}
