% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneutils.R
\name{insee_utils_sirene_aide}
\alias{insee_utils_sirene_aide}
\title{insee_utils_sirene_aide}
\usage{
insee_utils_sirene_aide()
}
\value{
un message d'information.
}
\description{
Retourne un message d'aide
}
\examples{
insee_utils_sirene_aide()
}
