% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneutils.R
\name{insee_utils_sirene_varenv_check}
\alias{insee_utils_sirene_varenv_check}
\title{insee_utils_sirene_varenv_check}
\usage{
insee_utils_sirene_varenv_check(var = "")
}
\arguments{
\item{var}{le nom de la variable d'environnement dont il faut verifier l'existence.}
}
\value{
un message d'avertissement si la variable n'est pas trouvee, sinon rien.
}
\description{
Retourne la valeur de la variable d'environnement INSEESIRENE_APIKEY
}
\examples{
\dontrun{
insee_utils_sirene_varenv_check("INSEESIRENE_APIKEY")
}
}
