% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneutils.R
\name{insee_utils_siret_multi_make}
\alias{insee_utils_siret_multi_make}
\title{insee_utils_siret_multi_make}
\usage{
insee_utils_siret_multi_make(multi = "")
}
\arguments{
\item{multi}{la chaine de caractere des criteres de la requete}
}
\value{
une chaine de caractere
}
\description{
Retourne l'url formatee d'une requete multicriteres
}
\examples{
\dontrun{
insee_utils_siret_multi_make("DEPCOM:132* & NAF:10.71C")
}
}
