% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneutils.R
\name{insee_utils_siret_sel}
\alias{insee_utils_siret_sel}
\title{insee_utils_siret_sel}
\usage{
insee_utils_siret_sel(df, myindic = "uniteLegale", mycateg = NA)
}
\arguments{
\item{df}{le data.frame utilise}

\item{myindic}{l'indicateur (famille d'information)}

\item{mycateg}{l'indicateur recherche}
}
\value{
une chaine de caractere : la valeur de l'indicateur
}
\description{
Retourne les informations du data.frame parse
}
\examples{
insee_utils_siret_sel(mydf, mycateg = "activitePrincipaleUniteLegale")
}
