% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneutils.R
\name{insee_utils_template_path}
\alias{insee_utils_template_path}
\title{acceder au repertoire du template pour tdb siret_}
\usage{
insee_utils_template_path(pack = "apisirene", tempname = "siren")
}
\arguments{
\item{pack}{le nom du package}
}
\value{
une chaine de char
}
\description{
Retourne le chemin du template Rmarkdown pour dashboard siret
}
\examples{
insee_utils_template_path(pack = "apisirene")
}
