% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nafniv5}
\alias{nafniv5}
\title{Codes et libelle naf niveau 5}
\format{
A data frame with x rows and y variables:
\describe{
\item{Code}{les codes naf de niveau 5}
\item{Libellé}{les libelles naf de niveau 5}
}
}
\usage{
nafniv5
}
\description{
A dataset contenant les codes et libelle de la naf niveau 5
Les variables sont les suivantes
}
\keyword{datasets}
