% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sirentreff}
\alias{sirentreff}
\title{Codes et libelle des tranches d'effectif de l'API Sirene}
\format{
A data frame with x rows and y variables:
\describe{
\item{code}{les codes de tranche d'effectif}
\item{lib}{les libelles de tranche d'effectif}
}
}
\usage{
sirentreff
}
\description{
A dataset contenant les codes et libelle des tranches d'effectif
Les variables sont les suivantes
}
\keyword{datasets}
