% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneutils.R
\name{trymycjlib}
\alias{trymycjlib}
\title{trymycjlib}
\usage{
trymycjlib()
}
\value{
Un tibble ou rien si pas d'echo
}
\description{
Retourne le code parmi un libelle de categorie juridique saisi par l'utilisateur.
}
\examples{
\dontrun{
trymycjlib()
}
}
