% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneutils.R
\name{trymynaflib}
\alias{trymynaflib}
\title{trymynaflib}
\usage{
trymynaflib()
}
\value{
Une chaine de caracteres ou rien si pas d'echos
}
\description{
Retourne le code parmi un libelle saisi par l'utilisateur.
}
\examples{
\dontrun{
trymynaflib()
}
}
