% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneparse.R
\name{insee_parse_siren_resp}
\alias{insee_parse_siren_resp}
\title{insee_parse_siren_resp}
\usage{
insee_parse_siren_resp(x)
}
\arguments{
\item{x}{une reponse siren mono unite legale}
}
\value{
un data.frame parse d'une requete siren mono unite legale
}
\description{
Parse le resultat d'une requete siren mono UL
}
\examples{
\dontrun{
insee_parse_siren_resp(objetnonparsedunerequetesiren)
}
}
