% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisireneparse.R
\name{insee_parse_siret_resp}
\alias{insee_parse_siret_resp}
\title{insee_parse_siret_resp}
\usage{
insee_parse_siret_resp(x)
}
\arguments{
\item{x}{une réponse siret sur un etablissement}
}
\value{
un data.frame parse d'une requête mono etab siret
}
\description{
Parse le résultat d'une requête sur UN numero siret
}
\examples{
insee_parse_siret_resp(objetnonparsedunerequetesiretmonoetab)
}
